(PC-23769)[PRO] refactor: split individual/educational code in OfferType
